# Use refpool optimized method for integer grouping