@bkamins 2021-01-24
Looks very nice. We only need to make sure we correctly disable trying this path in corner cases (like `BigInt` vector that has very large values, but with small range) as otherwise we will have overflow.