@bkamins 2021-01-26
I would also add a test that on 32-bit machine uses slow path and on 64-bit machine uses fast path (as `Int` is defined differently on both) and another test that on 64-bit machine uses slow path because the values are too large.