@nalimilan 2021-01-26
I've added checks and tests for overflows. I've also bumped into an already existing bug with `skipmissing=true` when there are only missing values so I've added more tests for that.