@bkamins 2021-01-31
Looks good. Thank you. Can you just please make sure that in the points I have just raised we are clear what happens?